style(render): fix floating-point arithmetic formatting

+++ sources/render/render_strip.c
@@ -6,7 +6,7 @@
 /*   By: peda-cos <peda-cos@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/28 20:53:39 by peda-cos          #+#    #+#             */
-/*   Updated: 2025/10/06 03:09:07 by peda-cos         ###   ########.fr       */
+/*   Updated: 2025/10/06 04:14:59 by peda-cos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,8 +81,8 @@ static void	draw_strip(t_engine *eng, t_ray *ray, int *rng, mlx_texture_t *tex)
 		texture_height = TEXTURE_CLAMP_ONE;
 	pixel.shift = get_door_texture_offset(eng, ray->map_x, ray->map_y);
 	step = (double)tex->height / (double)texture_height;
-	pixel.position = (double)(rng[0] + texture_height / CENTER_FACTOR
-			- (int)get_center_y(eng)) * step;
+	pixel.position = ((double)(rng[0] + texture_height / CENTER_FACTOR)
+			- get_center_y(eng)) * step;
 	y = rng[0];
 	while (y <= rng[1])
 	{
